move cache config
