+++ fastdeploy/model_executor/layers/attention/append_attn_backend.py
@@ -95,7 +95,7 @@ class AppendAttentionBackend(AttentionBackend):
         """
         super().__init__()
         self.attention_metadata: AppendAttentionMetadata = None
-        self.block_size: int = fd_config.parallel_config.block_size
+        self.block_size: int = fd_config.cache_config.block_size
         self.max_seq_len: int = fd_config.parallel_config.max_model_len
         self.rope_theta: float = (
             10000.0 if fd_config.model_config.rope_theta is None else fd_config.model_config.rope_theta
