+++ fastdeploy/model_executor/layers/attention/mla_attention_backend.py
@@ -113,7 +113,7 @@ class MLAAttentionBackend(AttentionBackend):
         self.attention_metadata: MLAAttentionMetadata = None
 
         # 基础配置
-        self.block_size: int = fd_config.parallel_config.block_size
+        self.block_size: int = fd_config.cache_config.block_size
         self.max_seq_len: int = fd_config.parallel_config.max_model_len
         self.rope_theta: float = (
             10000.0 if fd_config.model_config.rope_theta is None else fd_config.model_config.rope_theta
