+++ fastdeploy/worker/model_runner_base.py
@@ -44,6 +44,7 @@ class ModelRunnerBase(ABC):
         self.parallel_config = fd_config.parallel_config
         self.graph_opt_config = fd_config.graph_opt_config
         self.quant_config = fd_config.quant_config
+        self.cache_config = fd_config.cache_config
         # ... config
 
         self.device = device
