+++ fastdeploy/worker/worker_base.py
@@ -50,6 +50,7 @@ class WorkerBase(ABC):
         self.load_config = fd_config.load_config
         self.parallel_config = fd_config.parallel_config
         self.device_config = fd_config.device_config
+        self.cache_config = fd_config.cache_config
         # ... config
 
         # Device and Runner
